WL-0MLZVRB3501I5NSU: add fallback search tests, extract from fts tests, and add CLI needs-producer-review yes/no parsing tests
